+++ android/src/main/kotlin/io/adrop/adrop_ads/interstitial/FlutterAdropInterstitialAd.kt
@@ -87,7 +87,8 @@ class FlutterAdropInterstitialAd(
         return mapOf(
             "creativeId" to ad.creativeId,
             "txId" to ad.txId,
-            "campaignId" to ad.campaignId
+            "campaignId" to ad.campaignId,
+            "browserTarget" to ad.browserTarget
         )
     }
 }