+++ example/lib/views/interstitial_example.dart
@@ -51,7 +51,7 @@ class _InterstitialExampleState extends State<InterstitialExample> {
           // Callback: Called when the ad is successfully loaded
           onAdReceived: (ad) {
             debugPrint(
-                "interstitialAd received $unitId, ${ad.creativeId} ${ad.txId} ${ad.campaignId}");
+                "interstitialAd received $unitId, ${ad.creativeId} ${ad.txId} ${ad.campaignId} browserTarget: ${ad.browserTarget}");
             setState(() {
               isLoaded = true;
               errorCode = null;
@@ -66,8 +66,9 @@ class _InterstitialExampleState extends State<InterstitialExample> {
             });
           },
           // Callback: Called when the ad is clicked
-          onAdClicked: (_) {
-            debugPrint("interstitialAd clicked $unitId");
+          onAdClicked: (ad) {
+            debugPrint(
+                "interstitialAd clicked $unitId browserTarget: ${ad.browserTarget}");
           },
           // Callback: Called when the full-screen ad is presented
           onAdDidPresentFullScreen: (_) {